backend/requirements.txt
# ── Core Framework ────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.5.2
aiofiles==24.1.0
httpx==0.27.2
python-dotenv==1.0.1
bcrypt==4.0.1

# ── Database ──────────────────────────────────────────
psycopg2-binary==2.9.9

# ── AI / LLM ──────────────────────────────────────────
langchain==0.3.14
langchain-core==0.3.29
langchain-community==0.3.14
langchain-groq==0.2.4
langchain-google-genai==2.0.8
langgraph==0.2.62
chromadb==0.5.23
langchain-chroma==0.2.2
groq==0.13.0

# ── Google Gemini (Vision OCR) ────────────────────────
google-generativeai==0.8.3

# ── Image Processing ──────────────────────────────────
Pillow==11.0.0
PyMuPDF==1.24.14
opencv-python-headless==4.10.0.84
numpy==1.26.4

# ── Export ────────────────────────────────────────────
reportlab==4.2.2

# ── Scraping (for populate_topper_db.py pipeline) ─────
beautifulsoup4==4.12.3
lxml==5.3.0

# ── Testing ───────────────────────────────────────────
pytest==8.3.3
pytest-asyncio==0.24.0
